Includes jQuery, header and bigfoot in all pages
Since the script to auto-hide the navigation bar depends on jQuery
and needs to be present in all pages, we bundle it together with
bigfoot (which also depends on jQuery) in a single minified file
and link it in all pages.

This has the nice side-effect of making footnotes available in all
pages, which is something I've been wanting to do for a while.